Create
Create a program that reads the measurements of a wall (width and height). There is a type of paint to be used to paint the wall that yields 0.5 square meters per liter of paint, which costs $1785 per liter. Create a program that prompts the user for the measurements of the wall and displays the cost of painting the wall.